// HelpDeskFE/HelpDeskFETp/src/api/model/getApiMaintenancesParams.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export type GetApiMaintenancesParams = {
IdMaintenanceType?: number;
IdArea?: number;
IdDevice?: number;
DateFrom?: string;
DateTo?: string;
Keyword?: string;
PageNumber?: number;
PageSize?: number;
};